refactor: adjusting main view && improving component CollapsingToolbarScaffoldState

+++ app/src/main/java/br/com/raqfc/movieapp/common/presentation/composables/scaffold/BaseScaffold.kt
@@ -90,7 +90,7 @@ fun BaseScaffold(
                     actions = {
                         ActionMenu(
                             items = actions,
-                            viewWidth = actionsRowWidth
+                            viewWidth = actionsRowWidth.toFloat()
                         )
                     }
                 )
